requirements.txt
speechmatics-python
openai
gtts